--- TikTokDataforCreators/mac.env
@@ -1,1 +1,0 @@
-SYSTEM=MAC
