+++ app/harvest.py
@@ -2,11 +2,11 @@ from tiktokapipy.api import TikTokAPI
 import os
 import datetime
 import config
-import logging
 import tempfile
 import boto3
 import json
 import extract
+from vidvault_utils import timeit
 
 # Adjust this to capture more videos
 # API works by scrolling down page on TikTok,
@@ -15,36 +15,39 @@ import extract
 def get_scroll_time(user: str):
     filename = f'{user}'
     with TikTokAPI(scroll_down_time=1,navigation_retries=5, navigation_timeout=0, 
-                    data_dump_file=filename, headless=True, navigator_type='firefox') as api:
-        try:
-            user_object = api.user(user, video_limit=0)
-        except:
-            return 200
+                    data_dump_file=filename) as api:
+        user_object = api.user(user, video_limit=0)
     filename2 = f'{user}.UserResponse.json'
     with open(filename2, 'r') as file:
         json_data = json.load(file)
     video_count = json_data['UserModule']['stats'][user]['videoCount']
-    scroll_time = video_count / 2
-    return scroll_time
-        
+    os.remove(filename2)
+    
+    # Limit scroll time to within thsee ranges
+    
+    if video_count > config.HARVEST_SCROLL_TIME['MAX']:
+        return config.HARVEST_SCROLL_TIME['MAX']
+    
+    elif video_count < config.HARVEST_SCROLL_TIME['MIN']:
+        return config.HARVEST_SCROLL_TIME['MIN']
+    else:
+        return video_count     
 
-
+@timeit(message='Scrape TikTok')
 def run(user: str, 
         date: datetime.datetime) -> bool:
     """Takes in username and cleans relevant data for that user."""
     scroll_time = get_scroll_time(user=user)
-    print('Scroll time: ' +str(scroll_time))
     date_string = date.strftime('%m-%d-%Y')
     with tempfile.TemporaryDirectory(dir=config.LOCAL_PATH_PREFIX) as tmpdirname:
         filename = tmpdirname + f'/{date_string}'
         with TikTokAPI(scroll_down_time=scroll_time,navigation_retries=5, navigation_timeout=0, 
-                    data_dump_file=filename, headless=True, navigator_type='firefox') as api:
+                    data_dump_file=filename) as api:
             try:
                 user_object = api.user(user, video_limit=0)
                 upload_to_s3(directory=tmpdirname,
                                  user=user,
                                  date=date)
-                
                 return True
             except Exception as e:
                 print(e)
@@ -69,5 +72,7 @@ def upload_to_s3(directory: str,
 if __name__ == '__main__':
     user = 'tylerandhistummy'
     date = datetime.datetime.now()
-    run(user=user,
+    run(user='tytheproductguy',
         date=datetime.datetime.now())
+    extract.run(user=user,
+                date=date)
