+++ app_old/extract.py
@@ -5,20 +5,7 @@ import os
 import config
 import boto3
 import airtable_utils
-import pipeline
-from vidvault_utils import timeit
 
-def get_user_threshold(airtable_row_id: str):
-    """Implement dynamic thershold. All users default to 0.99
-    to start. Extract will check how many videos we scraped from harvest and 
-    compare to the threshold. videos_scraped / videos_scraped_threshold. If this ratio is less
-    than the threshold, we will descrease threshold and rerun"""
-    table = airtable_utils.get_table_data()
-    df = airtable_utils.convert_to_dataframe(airtable_table=table)
-    df = df.set_index('airtable_row_id')
-    threshold = df.loc[airtable_row_id, 'videos_scraped_threshold']
-    return threshold
-    
 def data_store_dict_return() -> dict:
     
     data_store_dict = {'video_id': [],
@@ -76,7 +63,8 @@ def extract_extras(data: dict, user: str):
             all_videos.append(df_page)
         except Exception as e:
             print(f'Extras page of harvest data skipped for user {user}')
-            pass
+            continue
+        
     df = pd.concat(all_videos)
     return df
 
@@ -95,7 +83,11 @@ def check_video_count(df: pd.DataFrame,
     """
     videos_scraped = len(df['video_id'].unique())
     videos_expected = get_user_video_count(data=data, user=user)
-    user_threshold = get_user_threshold(airtable_row_id=airtable_row_id)
+    
+    if videos_scraped / videos_expected < 0.99: # Scraper doesnt always work, set threshold to 95% or it will fail and rerun
+        assert 1 == 0
+    
+    
     
     airtable_utils.update_database_cell(row_id=airtable_row_id,
                                         field='videos_scraped',
@@ -104,20 +96,11 @@ def check_video_count(df: pd.DataFrame,
                                         field='total_videos',
                                         value=videos_expected)
     
-    if videos_scraped / videos_expected < user_threshold: # Scraper doesnt always work, set threshold to 95% or it will fail and rerun
-        new_user_threshold = user_threshold - config.VIDEOS_SCRAPED_THRESHOLD_DECREASE
-        airtable_utils.update_database_cell(row_id=airtable_row_id,
-                                            field='videos_scraped_threshold',
-                                            value=new_user_threshold)
-        print(f'{user} did not meet the videos scraped threshold, retrying!')
-        assert 1 == 0
-    
-    
     if videos_scraped == videos_expected:
         return True
 
     else:
-        print(f'{user} did not get all videos. {videos_scraped} / {videos_expected} scraped, but threshold met.')
+        print(f'{user} did not get all videos. {videos_scraped} / {videos_expected} scraped')
         return False
      
 def extract(data: dict, user: str, airtable_row_id: str) -> pd.DataFrame:
@@ -148,7 +131,7 @@ def extract(data: dict, user: str, airtable_row_id: str) -> pd.DataFrame:
     check_video_count(df=df, data=data, user=user, airtable_row_id=airtable_row_id)
     return df
 
-@timeit(message='Extract')
+
 def run(user: str,
         date: datetime.datetime,
         airtable_row_id: str):
@@ -167,7 +150,7 @@ def run(user: str,
     return True
 
 if __name__ == '__main__':
-    run(user='hi.surya',
+    run(user='tylerandhistummy',
         date=datetime.datetime.now(),
-        airtable_row_id='recLyU8bBKQl2tAJD')
+        airtable_row_id='recb4iqk60sDmE4cu')
