+++ app_old/pipeline.py
@@ -1,5 +1,7 @@
 import datetime
 import config
+import harvest
+import extract
 import download_videos
 import load
 import time
@@ -7,57 +9,51 @@ import aws_utils
 import airtable_utils
 from functools import wraps
 import time
-import math
-import vidvault_utils
+import pandas as pd
 
-TEST_USER = 'tytheproductguy'
-TEST_AIRTABLE_ROW = 'recKGvmEQKlfQ8600'
+TEST_USER = 'tylerandgerald'
+TEST_AIRTABLE_ROW = 'recMIbljDMr5dGaUQ'
 
-def check_if_this_is_pipeline_test(airtable_row_id: str):
-    table = airtable_utils.get_table_data()
-    df = airtable_utils.convert_to_dataframe(airtable_table=table)
-    df = df.set_index('airtable_row_id')
-    is_test = df.loc[airtable_row_id, 'test_run']
-    
-    if is_test == 'True':
-        return True
-    else: return False
+def timeit(func):
+    @wraps(func)
+    def timeit_wrapper(*args, **kwargs):
+        start_time = time.perf_counter()
+        result = func(*args, **kwargs)
+        end_time = time.perf_counter()
+        total_time = end_time - start_time
+        print(f'Pipeline took {total_time:.4f} seconds')
+        return result
+    return timeit_wrapper
 
-def check_if_threshold_set(airtable_row_id: str):
-    table = airtable_utils.get_table_data()
-    df = airtable_utils.convert_to_dataframe(airtable_table=table)
-    df = df.set_index('airtable_row_id')
-    threshold = df.loc[airtable_row_id, 'videos_scraped_threshold']
-    try:
-        threshold = float(threshold)
-    except:
-        return False
-    
-    if math.isnan(threshold):
-        return False
-    else: return True
-
-
+# TODO: have an input arg of user and then apply 
+@timeit
 def main(user_data: dict):
-
-    user_raw = user_data['user']
-    user = vidvault_utils.clean_user(user=user_raw)
+    
+    user = user_data['user']
     airtable_row_id = user_data['airtable_row_id']
     date=datetime.datetime.now()
-    start_time = time.time()
-    print(f'Pipeline for {user} started!')
-
-
+    
+    harvest.run(user=user, 
+                date=date)
+    print(f'Harvest for {user} complete')
+    
+    extract.run(user=user, 
+                date=date,
+                airtable_row_id=airtable_row_id)
+    print(f'Extract for {user} complete')
+    try:
+        load.run(user=user, 
+                 date=date)
+        print(f'Load for {user} complete')
+    except Exception as e:
+        print(e)
+        print(f'Load of analytics data failed on {user}, proceeding to video download')
+    
+    
     save_video = download_videos.run(user=user, 
                                      date=date)
     print(f'Videos saved for {user} complete')
     
-    if check_if_this_is_pipeline_test(airtable_row_id):
-        airtable_utils.update_database_cell(row_id=airtable_row_id,
-                                            field='test_run',
-                                            value="False")
-        assert 1 == 0, f"Test run for {user} complete!"
-        
     if save_video:
         try:
             url = aws_utils.create_presigned_url(bucket_name=config.BUCKET,
@@ -77,9 +73,7 @@ def main(user_data: dict):
                                                 value="True")
             print(e)
             print(f'Could not get url and update airtable for {user}')
-    elapsed_time = time.time() - start_time
-    total_time = round(elapsed_time, 4)
-    print(f'Pipeline for {user} | complete in {total_time:.4f} seconds')
+
 def run(user_data: dict):
     try:
         user = user_data['user']
@@ -89,7 +83,11 @@ def run(user_data: dict):
     except Exception as e:
         print(e)
         print(f'Pipeline failed on user {user}')
-        pass
+        bad_users_history = pd.read_csv(config.UserSignUpPath().bad_users)
+        bad_users_current_harvest = pd.DataFrame(data=[user_data])
+        bad_users_final = pd.concat([bad_users_history, bad_users_current_harvest])
+        bad_users_final.to_csv(config.UserSignUpPath().bad_users, index=False)
+        #airtable_utils.mark_video_upload_failed(row_id=airtable_row_id)  # Old update command
         airtable_utils.update_database_cell(row_id=airtable_row_id,
                                             field='upload_failed',
                                             value="True")
