+++ app_old/requirements.txt
@@ -5,7 +5,7 @@ aiosignal==1.3.1
 async-timeout==4.0.2
 attrs==22.2.0
 boto3
-botocore==1.27.59
+botocore
 certifi==2022.12.7
 charset-normalizer==3.1.0
 frozenlist==1.3.3
@@ -21,7 +21,6 @@ pyairtable==1.4.0
 pydantic==1.10.7
 pyee==9.0.4
 python-dateutil==2.8.2
-python-dotenv==1.0.0
 pytz==2022.7.1
 requests==2.28.2
 s3fs==2023.3.0
