+++ app_old/run.py
@@ -1,0 +1,1 @@
+scrape_completed
